requirements.txt
flask
flask_login
flask_wtf
flask_sqlalchemy
WTForms-Alchemy